style: rename module to match its content
BREAKING CHANGE: this invalidates user-side imports in old casing
