+++ src/LtHash.ts
