+++ src/index.ts
@@ -1,3 +1,6 @@
-import LtHash from "./lthash.ts"
+/* eslint-disable consistent-default-export-name/default-export-match-filename */
+
+import LtHash from "./LtHash.ts"
+
 export { LtHash }
 export default LtHash
